[docs] remove create and fund account
This document was a lot more like how to use a wallet and had little
to do with the developer journey. We need to work with the foundation
team to have a user onboarding guide that sits independent of this site
or another book called `use aptos` that walks folks through this.